requirements.txt
Django==1.7.2
django-braces==1.4.0
django-multiupload==0.3
django-sendfile==0.3.6
ipython==2.3.1
Pillow==2.6.1
simplejson==3.6.5
six==1.8.0
Unidecode==0.4.17
